+++ src/source.rs
@@ -1609,6 +1609,7 @@ impl Compiler {
             for slot in slots.iter().rev() {
                 state.emit(Op::SetLocal(*slot), &case.span);
             }
+            state.emit(Op::Pop, &case.span);
             let guard_next = if let Some(guard) = &case.guard {
                 self.expression(state, guard)?;
                 Some(state.jump_if_false(&case.span))